Optimize movegen: score-in-eighths and MULT_SMALL for multipliers
Performance optimizations for 68000:

1. Score in eighths: Store all scores pre-multiplied by 8 to eliminate
   the `score << 3` shift when computing equity. Required changing
   cross_scores from int8_t to int16_t (values can exceed 127).

2. MULT_SMALL macro: Avoid 70-cycle MULS instruction when multiplying
   by word_mult or letter_mult (always 1, 2, or 3). Uses simple adds:
   - ×1: no-op
   - ×2: val + val (4 cycles)
   - ×3: val + val + val (8 cycles)

3. Shadow loop optimization: Only loop for min(rack.total,
   num_unrestricted_multipliers) to skip zero-multiplier iterations.

Results (NWL23 10 games):
- Shadow: 103,396 → 100,324 frames (3.0% faster)
- NoShadow: 107,336 → 103,434 frames (3.6% faster)

+++ inc/kwg.h
@@ -155,7 +155,7 @@ int kwg_is_valid_word(const uint32_t *kwg, const MachineLetter *letters, int cou
 CrossSet compute_cross_set(const uint32_t *kwg,
                            const MachineLetter *prefix, int prefix_len,
                            const MachineLetter *suffix, int suffix_len,
-                           int8_t *cross_score);
+                           int16_t *cross_score);
 
 /*
  * Compute extension sets for a position given left and right tiles
